feat(Wallet) complete the filter API
It uses the current data only and doesn't extend with new types or
include new features in activity sources DBs.

Major changes:
- Partially filter by chain IDs
- Partially filter by Status if it is the case
- Partially filter by token types
- Filter by counterparty addresses
- Use wallet accounts for TO/FROM instead of filters

Closes: #10634

+++ services/wallet/transfer/database.go
@@ -279,7 +279,7 @@ func (db *Database) GetTransfersForIdentities(ctx context.Context, identities []
 	for _, identity := range identities {
 		subQuery := newSubQuery()
 		// TODO optimization: consider using tuples in sqlite and IN operator
-		subQuery = subQuery.FilterNetwork(identity.ChainID).FilterTransactionHash(identity.Hash).FilterAddress(identity.Address)
+		subQuery = subQuery.FilterNetwork(uint64(identity.ChainID)).FilterTransactionHash(identity.Hash).FilterAddress(identity.Address)
 		query.addSubQuery(subQuery, OrSeparator)
 	}
 	rows, err := db.client.QueryContext(ctx, query.String(), query.Args()...)
